requirements.txt
Django>=5.2.8,<6.0
django-cors-headers>=4.4.0,<5.0
whitenoise>=6.7.0,<7.0
gunicorn>=22.0.0,<23.0
python-dotenv>=1.0.1,<2.0
PyMySQL>=1.1.1,<2.0
cryptography>=42.0.0,<45.0
Pillow>=10.0.0,<11.0

